Small cleanup and reorganization of Shape .ts files
Imo, there's no reason why there should be really small .ts files defining stuff that could very well just... exist in another .ts file (per example, Small Duct Fans shouldn't have their own .ts file, but should be part of duct_fan.ts, same thing with all the Signs).

It'd also make things more organized (all Duct Fan instances in one file, all Sign instances in one file, all Bumper instances in one file).
