+++ src/ts/level.ts
@@ -3,16 +3,13 @@ import { Marble, bounceParticleOptions } from "./marble";
 import { Shape, SharedShapeData } from "./shape";
 import { MissionElementSimGroup, MissionElementType, MissionElementStaticShape, MissionElementItem, MisParser, MissionElementTrigger, MissionElementInteriorInstance, MissionElementTSStatic, MissionElementParticleEmitterNode, MissionElementSky } from "./parsing/mis_parser";
 import { StartPad } from "./shapes/start_pad";
-import { SignFinish } from "./shapes/sign_finish";
-import { SignPlain } from "./shapes/sign_plain";
 import { EndPad, fireworkSmoke, redSpark, redTrail, blueSpark, blueTrail } from "./shapes/end_pad";
 import { Gem } from "./shapes/gem";
 import { SuperJump, superJumpParticleOptions } from "./shapes/super_jump";
-import { SignCaution } from "./shapes/sign_caution";
 import { SuperBounce } from "./shapes/super_bounce";
-import { RoundBumper } from "./shapes/round_bumper";
+import { RoundBumper, TriangleBumper } from "./shapes/abstract_bumper";
 import { Helicopter } from "./shapes/helicopter";
-import { DuctFan } from "./shapes/duct_fan";
+import { DuctFan, SmallDuctFan } from "./shapes/duct_fan";
 import { AntiGravity } from "./shapes/anti_gravity";
 import { LandMine, landMineSmokeParticle, landMineSparksParticle } from "./shapes/land_mine";
 import { ShockAbsorber } from "./shapes/shock_absorber";
@@ -20,12 +17,10 @@ import { SuperSpeed, superSpeedParticleOptions } from "./shapes/super_speed";
 import { TimeTravel } from "./shapes/time_travel";
 import { Tornado } from "./shapes/tornado";
 import { TrapDoor } from "./shapes/trap_door";
-import { TriangleBumper } from "./shapes/triangle_bumper";
 import { Oilslick } from "./shapes/oilslick";
 import { Util, Scheduler } from "./util";
 import { PowerUp } from "./shapes/power_up";
 import { isPressed, releaseAllButtons, gamepadAxes, getPressedFlag, resetPressedFlag, hideTouchControls, maybeShowTouchControls, setTouchControlMode } from "./input";
-import { SmallDuctFan } from "./shapes/small_duct_fan";
 import { PathedInterior } from "./pathed_interior";
 import { Trigger } from "./triggers/trigger";
 import { InBoundsTrigger } from "./triggers/in_bounds_trigger";
@@ -39,7 +34,7 @@ import { Replay } from "./replay";
 import { Mission } from "./mission";
 import { PushButton } from "./shapes/push_button";
 import { state } from "./state";
-import { Sign } from "./shapes/sign";
+import { Sign, SignFinish, SignPlain, SignCaution } from "./shapes/sign";
 import { Magnet } from "./shapes/magnet";
 import { Nuke, nukeSmokeParticle, nukeSparksParticle } from "./shapes/nuke";
 import { TeleportTrigger } from "./triggers/teleport_trigger";
