+++ src/ts/shapes/sign.ts
@@ -22,3 +22,44 @@ export class Sign extends Shape {
 		}
 	}
 }
+
+/** A caution/danger sign. */
+export class SignCaution extends Shape {
+	dtsPath = "shapes/signs/cautionsign.dts";
+	shareMaterials = false;
+
+	constructor(element: MissionElementStaticShape) {
+		super();
+
+		// Determine the type of the sign
+		let type = element.datablock.slice("SignCaution".length).toLowerCase();
+		switch (type) {
+			case "caution": this.matNamesOverride["base.cautionsign"] = "caution.cautionsign"; break;
+			case "danger": this.matNamesOverride["base.cautionsign"] = "danger.cautionsign"; break;
+		}
+	}
+}
+
+/** The flickering finish sign, usually above the finish pad. */
+export class SignFinish extends Shape {
+	dtsPath = "shapes/signs/finishlinesign.dts";
+}
+
+/** A plain sign showing a direction, used in Marble Blast Gold. */
+export class SignPlain extends Shape {
+	dtsPath = "shapes/signs/plainsign.dts";
+	shareMaterials = false;
+
+	constructor(element: MissionElementStaticShape) {
+		super();
+
+		// Determine the direction to show
+		let direction = element.datablock.slice("SignPlain".length).toLowerCase();
+		switch (direction) {
+			case "right": this.matNamesOverride["base.plainsign"] = "right.plainsign"; break;
+			case "left": this.matNamesOverride["base.plainsign"] = "left.plainsign"; break;
+			case "up": this.matNamesOverride["base.plainsign"] = "up.plainsign"; break;
+			case "down": this.matNamesOverride["base.plainsign"] = "down.plainsign"; break;
+		}
+	}
+}
