--- src/ts/shapes/sign_finish.ts
@@ -1,6 +1,0 @@
-import { Shape } from "../shape";
-
-/** The flickering finish sign, usually above the finish pad. */
-export class SignFinish extends Shape {
-	dtsPath = "shapes/signs/finishlinesign.dts";
-}
