--- src/ts/shapes/triangle_bumper.ts
@@ -1,7 +1,0 @@
-import { AbstractBumper } from "./abstract_bumper";
-
-/** A triangle-shaped bumper. */
-export class TriangleBumper extends AbstractBumper {
-	dtsPath = "shapes/bumpers/pball_tri.dts";
-	sounds = ["bumper1.wav"];
-}
